# Improved GUI performance + fix #2493