--- a/Copy to SD Card root directory to update/Language Packs/language_hu.ini
+++ b/Copy to SD Card root directory to update/Language Packs/language_hu.ini
@@ -58,7 +58,7 @@ label_hotend_offset:Második fej eltolása
 label_stealth_chop:TMC StealthChop
 label_delta_configuration:Delta konfiguráció
 label_delta_tower_angle:Torony szögvágás
-label_delta_diagonal_rod:Diagonal Rod Trim
+label_delta_diagonal_rod:Átlós rúd igazítás
 label_delta_endstop:Végállás beállítások
 label_probe_offset:Szonda eltolás
 label_lin_advance:Lineáris haladás
@@ -133,7 +133,7 @@ label_ready:Kész
 label_busy:Feldolgozás alatt, kérlek várj...
 label_loading:Betöltés...
 label_unconnected:Nincs nyomtató csatlakoztatva!
-label_listening:TFT Hallgatási módban!
+label_listening:TFT figyelő módban!
 label_info:Infó
 label_invalid_value:Nincs érvényes érték megadva
 label_timeout_reached:Időkorlát elérve!
@@ -161,18 +161,18 @@ label_filament_runout:Szál kifutás!
 label_disable_steppers:Mot.tilt
 label_xy_unlock:X-Y Felold
 label_start_print:%s\nfájl kiválasztva.\nElindítod a nyomtatását?
-label_stop_print:Megállítod a nyomtatást?
+label_stop_print:Leállítod a nyomtatást?
 label_is_pause:Nyomtatás idején a művelet tiltott.\nSzünetelteted a nyomtatást?
 label_m0_pause:Szünetelés M0 paranccsal
 label_test:Teszt
-label_deploy:Bevetés
-label_stow:Elhelyezés
+label_deploy:Kilökés
+label_stow:Behúzás
 label_repeat:Ismétlés
 label_hs_on:HS: Be
 label_hs_off:HS: Ki
 label_nozzle:Fej
 label_bed:Ágy
-label_chamber:Kamra
+label_chamber:Gépház
 label_fan:Ventilátor
 label_bltouch:BLTouch
 label_touchmi:TouchMi
@@ -190,12 +190,12 @@ label_1_percent:1%
 label_5_percent:5%
 label_10_percent:10%
 label_percent_value:%d%%
-label_5_seconds:5 Mp.
-label_10_seconds:10 Mp.
-label_30_seconds:30 Mp.
-label_60_seconds:1 Perc.
-label_120_seconds:2 Perc.
-label_300_seconds:5 Perc.
+label_5_seconds:5 mp.
+label_10_seconds:10 mp.
+label_30_seconds:30 mp.
+label_60_seconds:1 perc
+label_120_seconds:2 perc
+label_300_seconds:5 perc
 label_white:Fehér
 label_black:Fekete
 label_blue:Kék
@@ -221,15 +221,15 @@ label_print:Nyomtatás
 label_extrude:Adagolás
 label_settings:Opciók
 label_screen_settings:Kijelző
-label_ui_settings:UI
+label_ui_settings:Felület
 label_sound:Hangok
 label_marlin_mode_settings:Marlin mód
 label_machine_settings:Gép
 label_parameter_settings:Tényezők
 label_feature_settings:Funkciók
 label_connection_settings:Kapcsolat
-label_serial_ports:S. Portok
-label_baudrate:Kapcsolat
+label_serial_ports:S. portok
+label_baudrate:Baud-érték
 label_eeprom_settings:EEPROM
 label_rgb_settings:LED Szín
 label_rgb_off:LED Ki
@@ -316,7 +316,7 @@ label_mesh_tuner:Háló hangoló
 label_case_light:Gépház világítás
 label_load_started:A szál betöltése folyamatban van,\nnem lehetséges további művelet, míg véget nem ér.\nKérlek várj!
 label_unload_started:A szál kiadása folyamatban van,\nnem lehetséges további művelet, míg véget nem ér.\nKérlek várj!
-label_heaters_on:Néhány fűtőelem még üzemel.\nKikapcsolod a fűtőelemeket?
+label_heaters_on:Néhány fűtőelem még üzemel.\nKikapcsolod a őket?
 label_print_finished:   Nyomtatás\n   elkészült!
 label_main_screen:Főoldal
 label_previous_print_data:Összegzés