… Copy to SD Card root directory to update/THEME_The Round Miracle Menu Material theme/TFT28/bmp/status_speed.bmp
… README.md
… TFT/src/User/API/Language/language_hu.h
… TFT/src/User/API/UI/ui_draw.h
… TFT/src/User/API/coordinate.c
… TFT/src/User/API/coordinate.h
… TFT/src/User/API/interfaceCmd.c
… TFT/src/User/API/menu.c
… TFT/src/User/API/parseACK.c
… TFT/src/User/Menu/MBL.c
… TFT/src/User/Menu/MeshTuner.c
… TFT/src/User/Menu/PrintingMenu.c
… TFT/src/User/Menu/Terminal.c
… TFT/src/User/Menu/TuneExtruder.c
… TFT/src/User/Menu/ZOffset.c
… TFT/src/User/config.ini
… TFT/src/User/os_timer.c