**IMPROVEMENTS:**
* Improved performance on GUI_DispOne() function: With the exception of text requiring transparency (GUI_TEXTMODE_TRANS), each character in a text is now displayed in a char by char mode (an LCD window is used) instead of pixel by pixel mode (an LCD window made by only one pixel). Below, some videos showing the difference with the default 50ms and a custom 20ms of refresh rate for scrolling text. The stuttering in the scrolling text (visible when the info in the menu are toggled every 2 seconds) present in the old code is now not present even at 20ms (it means whole menu is always updated within 20ms while in the old code it was required about 75ms to have no stuttering). Results are based on STM32F10X MCU. The improved performance are present in any menu displaying text (e.g. Printing menu) and on any TFT variant.
* Minor cleanup

fixes #2493

**PR STATE:** Ready for merge.

**BEFORE:**

**50ms refresh rate for scrolling text:**

https://user-images.githubusercontent.com/64427768/161955070-78e1a096-ea97-4d36-bca7-75dda4b0efff.mp4

**20ms refresh rate for scrolling text:**

https://user-images.githubusercontent.com/64427768/162009072-0a10a6b4-d944-438e-8aeb-d64752793f0e.mp4

**AFTER:**

**50ms refresh rate for scrolling text:**

https://user-images.githubusercontent.com/64427768/161955146-46da89e0-7f25-4486-8931-376eae66650b.mp4

**20ms refresh rate for scrolling text:**

https://user-images.githubusercontent.com/64427768/162007785-4a45bd25-3d75-43bb-ae0e-119709031abe.mp4