@kisslorand 2022-04-06
What should we look for? I see flickering in both videos...